refactor(notifications): move sync into it's own file
